tuple: include tuple data_offset into byte array
Tuple data offset was store in 2 byte uint16_t field. It is not always
needed and now we will be able to use the byte array instead of fixed
field, thus consuming only 1 byte in case of "tiny tuples". In the same
time the size of basic tuple became now 11 byte instead of 10. This is
going to be fixed.

Part of #5385